[refactor] #34 - 패키지 승인 / 설치 로직 분리
